통합 테스트(facade) 패키지 위치 수정\
* domain -> application
